requirements.txt
<<<<<<< HEAD
Django==1.7.5
django-debug-toolbar==1.2.2
djangorestframework==3.0.5
MySQL-python==1.2.5
sqlparse==0.1.14

=======
Django==1.8
django-debug-toolbar==1.3.0
djangorestframework==3.1.1
MySQL-python==1.2.5
simplejson==3.6.5
sqlparse==0.1.15
>>>>>>> baf5fae7f2c6fadbd2b58bd10d80d50338404c95
